refactor: remove glass folder
