+++ app-shared/src/commonMain/kotlin/io/github/droidkaigi/confsched/component/KaigiNavigationScaffold.kt
@@ -22,7 +22,6 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import dev.chrisbanes.haze.HazeState
 import dev.chrisbanes.haze.rememberHazeState
-import io.github.droidkaigi.confsched.component.glass.GlassLikeBottomNavigationBar
 import io.github.droidkaigi.confsched.droidkaigiui.KaigiPreviewContainer
 import io.github.droidkaigi.confsched.droidkaigiui.compositionlocal.LocalBottomNavigationBarsPadding
 import io.github.droidkaigi.confsched.droidkaigiui.extension.Empty
